+++ packages/fighting-design/image-preview/src/image-preview.vue
@@ -1,9 +1,11 @@
 <script lang="ts" setup name="FImagePreview">
   import { Props, Emits } from './image-preview'
-  import { ref } from 'vue'
+  import { ref, watch } from 'vue'
+  import { FButton } from '../../button'
   import { FToolbar } from '../../toolbar'
   import { FToolbarItem } from '../../toolbar-item'
   import { keepDecimal } from '../../_utils'
+  import { FPopup } from '../../popup'
   import {
     FIconChevronLeftVue,
     FIconChevronRightVue,
@@ -19,21 +21,42 @@
     ImagePreviewSwitchImageInterface as a,
     ImagePreviewOptionClickInterface as b,
     ImagePreviewOnImgMousewheelInterface as c,
-    ImagePreviewOptionClickTargetKey as e,
     OptionFunInterface as g
   } from './interface'
-  import type {
-    OrdinaryFunctionInterface as f,
-    HandleEventInterface as d
-  } from '../../_interface'
+  import type { OrdinaryFunctionInterface as f } from '../../_interface'
+  import type { ToolbarClickEmitInterface } from '../../toolbar/src/interface'
 
   const prop = defineProps(Props)
   const emit = defineEmits(Emits)
 
+  const isVisible: Ref<boolean> = ref<boolean>(prop.visible)
   const scale: Ref<number> = ref<number>(1)
   const rotate: Ref<number> = ref<number>(0)
   const previewShowIndex: Ref<number> = ref<number>(
     prop.showIndex > prop.imgList.length - 1 ? 0 : prop.showIndex
+  )
+
+  // 关闭图片预览
+  const handleClose: f = (): void => {
+    emit('update:visible', false)
+    prop.close && prop.close()
+  }
+
+  watch(
+    (): boolean => isVisible.value,
+    (newVal: boolean): void => {
+      // 监视 isVisible，如果变为假，则关闭
+      if (!newVal) {
+        handleClose()
+      }
+    }
+  )
+
+  watch(
+    (): boolean => prop.visible,
+    (newVal: boolean): void => {
+      isVisible.value = newVal
+    }
   )
 
   // 图片加载
@@ -42,6 +65,7 @@
 
     imgList.forEach((item: string): void => {
       const img: HTMLImageElement = new Image() as HTMLImageElement
+      console.log(img)
       img.src = item
     })
   }
@@ -60,24 +84,6 @@
       return
     }
     scale.value += 0.2
-  }
-
-  // 加载图片
-  const onEnter: f = (): void => {
-    imagPreload()
-  }
-
-  // 关闭图片预览
-  const handleClose: d = (evt: MouseEvent): void => {
-    emit('close', evt)
-    emit('update:visible', false)
-  }
-
-  // 点击遮罩层关闭
-  const packingClose: d = (evt: MouseEvent): void => {
-    if (prop.modalClose) {
-      handleClose(evt)
-    }
   }
 
   // 滚轮缩放
@@ -117,36 +123,34 @@
       }
     } as const
 
-    optionFun[type]()
+    if (optionFun[type]) {
+      optionFun[type]()
+    }
   }
 
   // 点击操作栏
-  const optionClick: b = ({ key }: e): void => {
+  const optionClick: b = (target: ToolbarClickEmitInterface): void => {
     const optionFun = {
-      '1': (): void => smaller(),
-      '2': (): void => bigger(),
-      '3': (): void => recovery(),
-      '4': (): void => {
+      1: (): void => smaller(),
+      2: (): void => bigger(),
+      3: (): void => recovery(),
+      4: (): void => {
         rotate.value += 90
       },
-      '5': (): void => {
+      5: (): void => {
         rotate.value -= 90
       }
     } as g
 
-    optionFun[key]()
+    if (optionFun[target.key as string]) {
+      optionFun[target.key as string]()
+    }
   }
 </script>
 
 <template>
-  <transition name="f-image-preview" @enter="onEnter">
-    <div
-      v-if="visible"
-      class="f-image-preview"
-      :style="{ zIndex }"
-      @click.self="packingClose"
-      @mousewheel="onImgMousewheel"
-    >
+  <div class="f-image-preview" @mousewheel="onImgMousewheel">
+    <f-popup v-model:visible="isVisible" :open="imagPreload">
       <img
         class="f-image-preview__exhibition"
         draggable="false"
@@ -159,39 +163,33 @@
 
       <!-- 左右切换按钮 -->
       <template v-if="imgList.length > 1">
-        <f-toolbar
-          v-if="imgList.length > 1"
-          class="right-button"
-          round
+        <f-button
+          class="f-image-preview__next"
+          circle
+          :before-icon="FIconChevronRightVue"
           @click="switchImage('next')"
-        >
-          <f-toolbar-item :icon="FIconChevronRightVue" icon-size="25px" />
-        </f-toolbar>
+        />
 
-        <f-toolbar
-          v-if="imgList.length > 1"
-          class="left-button"
-          round
+        <f-button
+          class="f-image-preview__prev"
+          circle
+          :before-icon="FIconChevronLeftVue"
           @click="switchImage('prev')"
-        >
-          <f-toolbar-item :icon="FIconChevronLeftVue" icon-size="25px" />
-        </f-toolbar>
+        />
       </template>
 
       <!-- 关闭按钮 -->
-      <f-toolbar
-        v-if="isCloseBtn"
-        class="close-button"
-        round
+      <f-button
+        class="f-image-preview__close"
+        circle
+        :before-icon="FIconCrossVue"
         @click="handleClose"
-      >
-        <f-toolbar-item :icon="FIconCrossVue" icon-size="20px" />
-      </f-toolbar>
+      />
 
       <!-- 操作栏 -->
       <f-toolbar
         v-if="isOption"
-        class="option-toolbar"
+        class="f-image-preview__option"
         round
         @click="optionClick"
       >
@@ -201,6 +199,6 @@
         <f-toolbar-item :icon="FIconRotateClockwiseVue" :data-key="4" />
         <f-toolbar-item :icon="FIconRotateAntiClockwiseVue" :data-key="5" />
       </f-toolbar>
-    </div>
-  </transition>
+    </f-popup>
+  </div>
 </template>