+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - 新增 `f-swap` 组件
 - 修复 `f-image-preview` 组件预览旋转方向错误的问题
+- 优化 `f-image-preview` 组件内部性能
+- `r-popup` 组件 `open` `open-end` `close` `close-end` 改为 `props` 参数
 
 ## 0.7.0-beta.2 (2022-10-22)